Added new ticket mechanism on conference page
