+++ shared/prop-types/ticket-type.js
@@ -1,0 +1,10 @@
+import { PropTypes } from 'react';
+export const ticketType = PropTypes.shape({
+  title: PropTypes.string,
+  releaseDate: PropTypes.shape({
+    iso: PropTypes.string,
+  }),
+  available: PropTypes.bool,
+  price: PropTypes.string,
+});
+
